+++ test/rules-1/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  plugins: ["custom-rules"],
+  extends: ["../../.eslintrc"],
+  rules: {
+    "custom-rules/test-rule": 1,
+  },
+};