+++ test/rules-1/custom-rules/index.js
@@ -1,0 +1,13 @@
+module.exports = {
+  rules: {
+    "test-rule": {
+      create: function (context) {
+        return {
+          TemplateLiteral(node) {
+            context.report(node, "template literals show test-rule error");
+          },
+        };
+      },
+    },
+  },
+};